Change signature to increase compatibility
As discussed here https://github.com/Ryochan7/DS4Windows/issues/140

+++ DS4WindowsApi/HidLibrary/NativeMethods.cs
@@ -392,6 +392,7 @@ namespace DS4Windows
 	    static internal extern int HidP_GetValueCaps(short reportType, ref byte valueCaps, ref short valueCapsLength, IntPtr preparsedData);
 
         [DllImport("hid.dll")]
-        static internal extern bool HidD_GetSerialNumberString(IntPtr HidDeviceObject, byte[] Buffer, ulong BufferLength);
+        static internal extern bool HidD_GetSerialNumberString(IntPtr HidDeviceObject, byte[] Buffer, uint BufferLength);
+        //static internal extern bool HidD_GetSerialNumberString(IntPtr HidDeviceObject, byte[] Buffer, ulong BufferLength);
     }
 }
